+++ src/main/java/seedu/address/logic/commands/IncrementCommand.java
@@ -58,9 +58,9 @@ public class IncrementCommand extends Command {
 
         Person personToEdit = lastShownList.get(index.getZeroBased());
 
-        int currentCount = lastShownList.get(index.getZeroBased()).getTransactionCount().getIntValue();
-        int incrementCount = editPersonDescriptor.getTransactionCount().get().getIntValue();
-        int finalAmount = currentCount + incrementCount;
+        long currentCount = lastShownList.get(index.getZeroBased()).getTransactionCount().getLongValue();
+        long incrementCount = editPersonDescriptor.getTransactionCount().get().getLongValue();
+        long finalAmount = currentCount + incrementCount;
 
         editPersonDescriptor.setTransactionCount(new TransactionCount(String.valueOf(finalAmount)));
